Change compiler config to not fork
This commit moves the module opening options to a jvm.config file,
which applies to the JVM running Maven itself, and avoids forking
a new JVM for compilation. This makes the build faster.

The toolchains plugin is thus removed in favor of the enforcer plugin
to ensure a minimum Java version for the maven process itself.

Signed-off-by: Eric Bottard <eric.bottard@broadcom.com>